moved etl folder, fixed missing get_raw_data
