+++ README.md
@@ -38,7 +38,7 @@ To install the hooks run
 
 `source venvlocal/bin/activate` (Adds the venv directory to your PATH so running `python` will use the venv python and the packages it installed)
 
-`pip install -r athena-scripts/raw_stage_optimisation_solution/scripts/requirements.txt` (Installs packages from [requirements.txt](athena-scripts/raw_stage_optimisation_solution/scripts/requirements.txt) - only needs to be run first time)
+`pip install -r athena-scripts/raw_stage_optimisation_solution/scripts/requirements.txt` (Installs packages from [requirements.txt](raw_to_stage/requirements.txt) - only needs to be run first time)
 
 After this, the hooks defined under the [.husky](.husky) directory will automatically run when you commit or push.&ast;
 The [lint-staged](https://github.com/okonet/lint-staged) library is used to only run certain tasks if certain files are modified.
