+++ raw_to_stage/raw_to_stage_etl/strategies/BackfillStrategy.py
@@ -2,8 +2,8 @@
 
 from ..util.database_utilities import (get_all_previous_processed_dts, get_all_processed_times_per_day,
                                        get_min_timestamp_from_previous_run)
+from ..util.DataPreprocessing import get_last_processed_time, get_penultimate_processed_dt
 from ..util.exceptions.UtilExceptions import OperationFailedException
-from ..util.pandas_dataframe_utilities import get_last_processed_time, get_penultimate_processed_dt
 from .Strategy import Strategy
